% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.microbiomeBetareg.R
\name{ds.microbiomeBetareg}
\alias{ds.microbiomeBetareg}
\title{Fits beta regression models for rates and proportions via maximum likelihood}
\usage{
ds.microbiomeBetareg(
  df = NULL,
  formula = NULL,
  na.action,
  weights,
  offset,
  link = c("logit", "probit", "cloglog", "cauchit", "log", "loglog"),
  link.phi = NULL,
  type = c("ML", "BC", "BR"),
  control = betareg.control(...),
  model = TRUE,
  y = TRUE,
  x = FALSE,
  datasources = NULL
)
}
\arguments{
\item{df}{is a string character for the data.frame object}

\item{formula}{symbolic description of the model.}

\item{na.action}{is a string character specifying how NA data shall be treated.}

\item{weights}{optional numeric vector of case weights.}

\item{offset}{optional numeric vector with an a priori known component to be included in the linear predictor for the mean. In betareg.fit, offset may also be a list of two offsets for the mean and precision equation, respectively.}

\item{link}{character specification of the link function in the mean model (mu). Currently, "logit", "probit", "cloglog", "cauchit", "log", "loglog" are supported. Alternatively, an object of class "link-glm" can be supplied.}

\item{link.phi}{character specification of the link function in the precision model (phi). Currently, "identity", "log", "sqrt" are supported. The default is "log" unless formula is of type y ~ x where the default is "identity" (for backward compatibility). Alternatively, an object of class "link-glm" can be supplied.}

\item{type}{character specification of the type of estimator. Currently, maximum likelihood ("ML"), ML with bias correction ("BC"), and ML with bias reduction ("BR") are supported.}

\item{control}{a list of control arguments specified via betareg.control.}

\item{model}{is a logical. If TRUE, the model frame from the model fit will be returned.}

\item{y}{is a logical. If TRUE, the model matrix from the model fit will be returned.}

\item{x}{is a logical. If TRUE, the response from the model fit will be returned.}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login. If the \code{datasources} argument is not specified the default set of connections will be used: see \code{\link{datashield.connections_default}}.}
}
\value{
\code{ds.microbiomeBetareg} returns a beta regression model
}
\description{
This function is similar to the native R function betareg from the betareg package
}
\details{
The function calls the server-side function \code{microbiomeBetaregDS} that computes the beta regression model
from data.frame. The function uses parametrization with mean and precision parameter.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
