% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.microbiomeMZILN.R
\name{ds.microbiomeMZILN}
\alias{ds.microbiomeMZILN}
\title{Computes the regression for microbiome analysis based on multivariate zero-inflated logistic normal model}
\usage{
ds.microbiomeMZILN(
  SumExp = NULL,
  taxa = NULL,
  covariates = NULL,
  datasources = NULL
)
}
\arguments{
\item{SumExp}{is a string character describing the SummarizedExperiment object}

\item{taxa}{is a string character for the microbiome variable denominator (can also be a vector of microbiome variables)}

\item{covariates}{is a string character of covariates to be examined along the microbiome variables (can also be a vector of covariates).}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
XXXXX
}
\description{
This function is similar to the native R function from the IFAA package
}
\details{
The function calls the server-side function \code{microbiomeMZILNDS} that computes the
regression from a SummarizedExperiment object. SummarizedExperiment objects can be computed using the \code{ds.summarizedExperiment} function.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
