% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.summarizedExperiment.R
\name{ds.summarizedExperiment}
\alias{ds.summarizedExperiment}
\title{Computes the SummarizedExperiment object}
\usage{
ds.summarizedExperiment(
  df = NULL,
  microbiomeData = NULL,
  covariateData = NULL,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{df}{is a string character of the data.frame}

\item{microbiomeData}{is a string character of a microbiome variable of interest (can also be a vector of microbiome variables)}

\item{covariateData}{is a string character of a covariate variable of interest (can also be a vector of covariates)}

\item{newobj}{is the name of the new object which is created with this function}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.summarizedExperiment} or default name \code{sumExp.newobj}
}
\description{
This function is similar to the native R function from Bioconductor
}
\details{
The function calls the server-side function \code{summarizedExperimentDS} that computes the
SummarizedExperiment from a data.frame and assigns the object to the server-side.
The new object is named by the user using the \code{newobj} argument, otherwise it is named \code{sumExp.newobj} by default.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
