% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.microbiomeIFAA.R
\name{ds.microbiomeIFAA}
\alias{ds.microbiomeIFAA}
\title{Computes the association of microbiome data with covariates}
\usage{
ds.microbiomeIFAA(
  SumExp = NULL,
  microbVar = "all",
  testCov = NULL,
  ctrlCov = NULL,
  sampleIDname = NULL,
  testMany = TRUE,
  ctrlMany = FALSE,
  nRef = 40,
  nRefMaxForEsti = 2,
  refTaxa = NULL,
  adjust_method = "BY",
  fdrRate = 0.15,
  paraJobs = NULL,
  bootB = 500,
  standardize = FALSE,
  sequentialRun = FALSE,
  refReadsThresh = 0.2,
  taxDropThresh = 0,
  SDThresh = 0.05,
  SDquantilThresh = 0,
  balanceCut = 0.2,
  verbose = TRUE,
  type = c("split", "pooled"),
  datasources = NULL
)
}
\arguments{
\item{SumExp}{is a string character describing the SummarizedExperiment object}

\item{microbVar}{This takes a single or vector of microbiome variable names (e.g., taxa, OTU and ASV names) of interest. Default is "all" meaning all microbiome variables will be analyzed. If a subset of microbiome variables is specified, the output will only contain the specified variables, and p-value adjustment for multiple testing will only be applied to the subset.}

\item{testCov}{is a string character of covariates to be examined along the microbiome variables (can also be a vector of covariates).}

\item{ctrlCov}{is a string character for the covariates that will be adjusted in the model (can also be a vector of microbiome variables)}

\item{sampleIDname}{is a string character for the sample ID variable.}

\item{testMany}{is a logical. If 'TRUE' and 'covariates' are set to NULL, then all variables in the 'covariates' will be used.}

\item{ctrlMany}{is a logical. If 'TRUE' and 'confounders' are set to NULL, then all variables except the 'coviariates' will be used as confounders.}

\item{nRef}{number of randomly picked reference taxa used in phase 1.}

\item{nRefMaxForEsti}{maximum number of final reference taxa used in phase 2.}

\item{refTaxa}{vector of taxa or OTU or ASV names. Theses are reference taxa specified by the user to be used in phase 1.}

\item{adjust_method}{The adjusting method for p value adjustment. Default is "BY" for dependent FDR adjustment. It can take any adjustment method for the p.adjust function in R.}

\item{fdrRate}{The false discovery rate for identifying taxa/OTU/ASV associated with 'covariates'.}

\item{paraJobs}{If 'sequentialRun' is FALSE, this specifies the number of parallel jobs that will be registered to run the algoithm. If specified as NULL, it will automatically detect the cores to decide the number of parallel jobs.}

\item{bootB}{number of bootstrap samples for obtaining confidence interval of estimates in phase 2 for the high dimensional regression. Default is 500.}

\item{standardize}{is a logical. If 'TRUE', the design matrix for X will be standardized in the analyses and the results. Default is FALSE.}

\item{sequentialRun}{is a logical. Defines whether there are parallel jobs or not.}

\item{refReadsThresh}{The threshold of proportion of non-zero sequencing reads for choosing the reference taxon in phase 2. Default is 0.2 meaning that at least 20\% non-zero sequencing reads are necessary.}

\item{taxDropThresh}{The threshold of number of non-zero sequencing reads for each taxon to be dropped from the analysis. Default is 0 which means that taxon without any sequencing reads will be dropped from the analysis.}

\item{SDquantilThresh}{The threshold of the quantile of standard deviation of sequencing reads, above which could be selected as a reference taxon. Default is 0.}

\item{balanceCut}{The threshold of the proportion of non-zero sequencing reads in each group of a binary variable for choosing the final reference taxa in phase 2. The default is 0.2 meaning at least 20\% non-zero sequencing reads in each group are needed to be eligible for being chosen as a final reference taxon.}

\item{verbose}{Whether the process message is printed out to the console. Default is TRUE.}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}

\item{SDTresh}{The threshold of standard deviations of sequencing reads for being chosen as the reference taxon in phase 2. The default is 0.05 which means the standard deviation of sequencing reads should be at least 0.05 in order to be chosen as a reference taxon.}
}
\value{
\code{ds.microbiomeIFAA} returns the association of the microbiome data with the covariates
}
\description{
This function is similar to the native R function from the IFAA package
}
\details{
The function calls the server-side function \code{microbiomeIFAADS} that computes the
association analysis from a SummarizedExperiment object. SummarizedExperiment objects can be computed using the \code{ds.summarizedExperiment} function.
}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
