% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.summarizedExperiment.R
\name{ds.summarizedExperiment}
\alias{ds.summarizedExperiment}
\title{Computes the SummarizedExperiment object}
\usage{
ds.summarizedExperiment(
  microbiomeData = NULL,
  covariateData = NULL,
  newobj = NULL,
  datasources = NULL
)
}
\arguments{
\item{microbiomeData}{is a string character for a data.frame containing the microbiome variables of interest}

\item{covariateData}{is a string character for a data.frame containing the covariates variables of interest}

\item{newobj}{is the name of the new object which is created with this function}

\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login}
}
\value{
the object specified by the \code{newobj} argument of \code{ds.summarizedExperiment} or default name \code{sumExp.newobj}
}
\description{
This function is similar to the native R function from Bioconductor
}
\details{
The function calls the server-side function \code{summarizedExperimentDS} that computes the
SummarizedExperiment from a data.frame and assigns the object to the server-side.
The new object is named by the user using the \code{newobj} argument, otherwise it is named \code{sumExp.newobj} by default.
}
\examples{
\dontrun{

  # connecting to the Opal servers

  require('DSI')
  require('DSOpal')
  require('dsBaseClient')
  require('dsIntestinalMicrobiomicsClient')

  builder <- DSI::newDSLoginBuilder()
  builder$append(server = "study1",
                 url = "http://192.168.56.100:8080/",
                 user = "administrator", password = "datashield_test&",
                 table = "MicrobSIM.MicrobSIM1", driver = "OpalDriver")
  builder$append(server = "study2",
                 url = "http://192.168.56.100:8080/",
                 user = "administrator", password = "datashield_test&",
                 table = "MicrobSIM.MicrobSIM2", driver = "OpalDriver")
  builder$append(server = "study3",
                 url = "http://192.168.56.100:8080/",
                 user = "administrator", password = "datashield_test&",
                 table = "MicrobSIM.MicrobSIM3", driver = "OpalDriver")
  logindata <- builder$build()

  connections <- DSI::datashield.login(logins = logindata, assign = TRUE, symbol = "D")


  # Create data.frames with microbiome and covariate data of interest

  ds.dataFrame(x = c("D$P_ACTINOBACTERIA",
                     "D$P_BACTEROIDETES",
                     "D$P_FIRMICUTES",
                     "D$P_VERRUCOMICROBIA"),
               newobj = "microbdata",
               stringsAsFactors = FALSE)

  ds.dataFrame(x = c("D$AGE",
                     "D$SEX",
                     "D$WEIGHT",
                     "D$HEIGHT"),
               newobj = "covdata",
               stringsAsFactors = FALSE)


  # Create the summarizedExperiment object on the server-side based on the microbiome and covariate data.frames

  ds.summarizedExperiment(microbiomeData = "microbdata",
                          covariateData = "covdata",
                          newobj = "SumExpT")

  # clear the Datashield R sessions and logout
  datashield.logout(connections)
}

}
\author{
Florian Schwarz for the German Institute of Human Nutrition
}
